requirements.txt
aiogram==2.23.1
Pillow==9.3.0
python-dotenv==0.21.0
